backend/app/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
pillow
python-multipart

# later you will add:


# NEW FOR AI
ultralytics
opencv-python
numpy
pydantic-settings

# ultralytics
# opencv-python
